+++ src/backend/player.cpp
@@ -12,7 +12,7 @@ namespace backend {
         constexpr double kStandingDistance = 0.1;
     }
 
-    Player::Player(uint64_t single_id, double x, double y, SingleStats& single_stats) {
+    Player::Player(uint32_t single_id, double x, double y, SingleStats& single_stats) {
         main_single_ = std::make_unique<Single>(single_id, x, y, single_stats);
     }
