+++ src/backend/single/single.cpp
@@ -19,7 +19,7 @@ namespace backend {
         constexpr double kOutOfReachSpeedMultiplier = 1.4;
     }
 
-    Single::Single(uint64_t id, double x, double y, const SingleStats& single_stats) : id_(id),
+    Single::Single(uint32_t id, double x, double y, const SingleStats& single_stats) : id_(id),
         x_(x), y_(y), single_state_(SingleState::STANDING), single_stats_(single_stats) {
         goal_x_ = x;
         goal_y_ = y;
@@ -108,7 +108,7 @@ namespace backend {
         return { x_, y_};
     }
 
-    const uint64_t Single::GetId() const {
+    const uint32_t Single::GetId() const {
         return id_;
     }
 }
