+++ src/backend/single/single.h
@@ -7,7 +7,7 @@
 namespace backend {
     class Single {
         public:
-            Single(uint64_t id, double x, double y, const SingleStats& single_stats_);
+            Single(uint32_t id, double x, double y, const SingleStats& single_stats_);
 
             // Update the intention of the single
             void UpdateIntention(double time_delta);
@@ -22,11 +22,11 @@ namespace backend {
             const Point GetPosition() const;
 
             // Get ID of the single
-            const uint64_t GetId() const;
+            const uint32_t GetId() const;
 
         private:
             // Unique ID
-            uint64_t id_;
+            uint32_t id_;
 
             // Switch to a different state
             void SwitchSingleState(SingleState state);
